can't get the followers palettes to work correctly for some species, so I disabled it in the mean time.
removed flickering of the rustboro lights

+++ src/field_player_avatar.c
@@ -1301,7 +1301,7 @@ u8 unref_GetRivalAvatarGenderByGraphicsId(u16 gfxId)
     }
 }
 
-u8 GetPlayerAvatarGenderByGraphicsId(u16 gfxId)
+u16 GetPlayerAvatarGenderByGraphicsId(u16 gfxId)
 {
     switch (gfxId)
     {
